requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
Pillow==10.0.0
opencv-python==4.8.0.76
requests==2.31.0
tqdm==4.66.1
plotly==5.16.1
